Fixes:  from f5.bigip.bigip import BigIP --> from f5.bigip import BigIP
Reference Issue:   https://github.com/F5Networks/f5-common-python/pull/27

Purpose:           We've decided to put base class definitions that match
directory names into the corresponding __init__.py

Changes:           mv bigip/bigip.py --> bigip/__init__.py
